finish book crud

+++ Services/BooksService.cs
@@ -1,4 +1,5 @@
-using System;
+using System.Collections.Generic;
+using System.Linq;
 using FindbookApi.Models;
 
 namespace FindbookApi.Services
@@ -12,9 +13,31 @@ namespace FindbookApi.Services
             db = context;
         }
 
+        public List<Book> All(BooksFilter filter)
+        {
+            return db.Books
+                .OrderBy(x => x.Id)
+                .Skip(filter.Skip)
+                .Take(filter.Take)
+                .ToList();
+        }
+
         public Book Add(Book book)
         {
+            if (book.Id > 0)
+                book.Id = 0;
+            if (IsDuplicate(book))
+                return null;
             db.Books.Add(book);
+            db.SaveChanges();
+            return book;
+        }
+
+        public Book Edit(Book book)
+        {
+            if (db.Books.Find(book.Id) == null)
+                throw new System.ArgumentException("ads");
+            db.Books.Update(book);
             db.SaveChanges();
             return book;
         }
@@ -22,8 +45,14 @@ namespace FindbookApi.Services
         public void Delete(int id)
         {
             Book book = db.Books.Find(id);
-            db.Books.Remove(book);
-            db.SaveChangesAsync();
+            if (book != null)
+            {
+                db.Books.Remove(book);
+                db.SaveChanges();
+            }
         }
+
+        private bool IsDuplicate(Book book) =>
+            db.Books.Any(b => b.Title == book.Title && b.Author == book.Author);
     }
 }
